refactor: 로또 오름차순의 검증을 IssuedLotto로 수정
